feat(Links): Link component
Create Link component and use it on header
