+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import './App.css';
-import Footer from './components/Footer';
-import Header from './components/Header';
+import Footer from './/components/footer/footer';
+import Header from './components/header/header';
 import { Home } from './components/Views';
 
 const App = () => {
