+++ src/components/common/Link.jsx
@@ -1,0 +1,9 @@
+import "./styles.css"
+
+const Link = ({name, label, active = false}) => {
+    return(
+        <a className={`navLink ${name} ${active ? "active" : ""}`}>{label}</a>
+    )
+}
+
+export default Link;
