+++ src/components/common/index.js
@@ -1,5 +1,7 @@
 import ToolTip from "./ToolTip";
+import Link from "./Link"
 
 export {
-    ToolTip
+    ToolTip,
+    Link
 }
