+++ src/components/common/styles.css
@@ -21,3 +21,12 @@
     border: 0.5px solid gray;
     font-size: 1rem;
 }
+
+
+/* Link */
+.navLink {
+    font-family: monospace;
+    color: rgb(96 123 150);
+    padding: 0 1.5rem;
+    border-right: 1px solid #1e2d3d;
+}
